Исправил замечания ДЗ 3 второй раз

+++ Minibank.Core/Domains/Accounts/Services/AccountService.cs
@@ -1,6 +1,7 @@
 ﻿using Minibank.Core.Domains.Accounts.Repositories;
 using Minibank.Core.Domains.MoneyTransfers;
 using Minibank.Core.Domains.MoneyTransfers.Services;
+using Minibank.Core.Domains.Users.Repositories;
 using Minibank.Core.Domains.Users.Services;
 using System;
 using System.Collections.Generic;
@@ -15,40 +16,50 @@ namespace Minibank.Core.Domains.Accounts.Services
     public class AccountService : IAccountService
     {
         private readonly IAccountRepository _accountRepository;
-        private readonly IUserService _userService;
+        private readonly IUserRepository _userRepository;
         private readonly ICurrencyСonversion _currencyСonversion;
         private readonly IMoneyTransferService _moneyTransferService;
 
-        public AccountService(IAccountRepository accountRepository, IUserService userService,
-            ICurrencyСonversion currencyConversion, IMoneyTransferService moneyTransferService)
+        public AccountService(IAccountRepository accountRepository, ICurrencyСonversion currencyConversion, 
+            IMoneyTransferService moneyTransferService, IUserRepository userRepository)
         {
             _accountRepository = accountRepository;
-            _userService = userService;
             _currencyСonversion = currencyConversion;
             _moneyTransferService = moneyTransferService;
+            _userRepository = userRepository;
         }
+
         public void Create(Account account)
         {
-            if (_userService.GetUser(account.UserId) == null)
+            if (!_userRepository.Exists(account.UserId))
             {
                 throw new ValidationException("Пользователя с таким id не существует");
             }
-
-
             if (!Enum.IsDefined(typeof(permittedCurrencies), account.Currency))     
             {
                 throw new ValidationException("Задана недопустюмая валюта счета");
             }
+
             _accountRepository.Create(account);
         }
 
         public void Delete(string id)
         {
+            if (!_accountRepository.Exists(id))
+            {
+                throw new ValidationException("Аккаунта с таким id не существует");
+            }
+
             _accountRepository.Delete(id);
         }
 
         public Account GetUserAccounts(string id)
         {
+            if (!_accountRepository.Exists(id))
+            {
+                throw new ValidationException("Аккаунта с таким id не существует");
+            }
+
             return _accountRepository.GetUserAccounts(id);
         }
 
@@ -59,59 +70,69 @@ namespace Minibank.Core.Domains.Accounts.Services
 
         public void Close(String id)
         {
-            if (_accountRepository.GetUserAccounts(id).AmoumtOnAccount != 0)
+            Account account = GetUserAccounts(id);
+            if (account == null)
+            {
+                throw new ValidationException("Аккаунта с таким id не существует");
+            }
+            if (account.AmoumtOnAccount != 0)
             {
                 throw new ValidationException("Нельзя закрыть аккаунт с деньгами на нем");
             }
+
             _accountRepository.CloseAccount(id);
         }
+
         public double CalculateCommission(double amount, string fromAccountId, string toAccountId)
         {
-
             var fromAccountUserId = _accountRepository.GetUserAccounts(fromAccountId).UserId;
             var toAccountUserId = _accountRepository.GetUserAccounts(toAccountId).UserId;
             if (fromAccountUserId == toAccountUserId)
             {
                 return 0.0;
             }
+
             double commission = amount * 0.02;
             return Math.Round(commission, 2);
         }
+
         public void TransferMoney(double amount, string fromAccountId, string toAccountId)
         {
+            Account fromAccount = _accountRepository.GetUserAccounts(fromAccountId);
+            Account toAccount = _accountRepository.GetUserAccounts(toAccountId);
             if (amount <= 0)
             {
                 throw new ValidationException("Неправильна введена сумма перевода");
             }
-            if (_accountRepository.GetUserAccounts(fromAccountId) == null)
+            if (fromAccount == null)
             {
                 throw new ValidationException("Неправильно введен id аккаунта отправителя");
             }
-            if (_accountRepository.GetUserAccounts(toAccountId) == null)
+            if (toAccount == null)
             {
                 throw new ValidationException("Неправильно введен id аккаунта получателя");
             }
-            if (!_accountRepository.GetUserAccounts(fromAccountId).IsOpen)
+            if (!fromAccount.IsOpen)
             {
                 throw new ValidationException("Аккаунт отправителя закрыт");
             }
-            if (!_accountRepository.GetUserAccounts(toAccountId).IsOpen)
+            if (!toAccount.IsOpen)
             {
                 throw new ValidationException("Аккаунт получателя закрыт");
             }
-            if (_accountRepository.GetUserAccounts(fromAccountId).AmoumtOnAccount < amount)
+            if (fromAccount.AmoumtOnAccount < amount)
             {
                 throw new ValidationException("Недостаточно средств");
             }
 
-            var fromAccountCurrency = _accountRepository.GetUserAccounts(fromAccountId).Currency;
-            var toAccountCurrency = _accountRepository.GetUserAccounts(toAccountId).Currency;
-
+            var fromAccountCurrency = fromAccount.Currency;
+            var toAccountCurrency = toAccount.Currency;
             double resultAmount = amount - CalculateCommission(amount, fromAccountId, toAccountId);
             if (fromAccountCurrency != toAccountCurrency)
             {
                 resultAmount = _currencyСonversion.Converting(amount, fromAccountCurrency, toAccountCurrency);
             }
+
             resultAmount = Math.Round(resultAmount, 2);
             _accountRepository.SubAmount(fromAccountId, amount);
             _accountRepository.AddAmount(toAccountId, resultAmount);
